QtyDeliveredFromOrderToInvoiceTest: junit5+mockito
